+++ Etape_2/quest-spring-boot/src/main/java/com/quest/etna/repositories/UserRepository.java
@@ -4,7 +4,7 @@ import com.quest.etna.model.User;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepository extends CrudRepository<User, Long> {
+public interface UserRepository extends CrudRepository<User, Integer> {
 
     @Query("SELECT u FROM User u WHERE u.username = :username")
     User getUserByUsername(String username);
